Regorganized Components, Queries & Mutations
Reorganized Components into folders. Combined queries and mutations.
