+++ frontend/components/User.js
@@ -1,6 +1,6 @@
 import { Query } from 'react-apollo';
 import PropTypes from 'prop-types';
-import { CURRENT_USER_QUERY } from '../queries/CurrentUser';
+import { CURRENT_USER_QUERY } from '../queries/User';
 
 const User = props => (
     <Query {...props} query={CURRENT_USER_QUERY}>
