+++ frontend/components/admin/meat/DeleteMeat.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { DELETE_MEAT_MUTATION } from '../../../mutations/Meat';
 import { ALL_MEATS_QUERY } from '../../../queries/Meat';
 import { ConfirmDialog } from '../../styles/ConfirmDialog';
-import { ErrorAlert } from '../../ErrorAlert';
+import { ErrorAlert } from '../../elements/ErrorAlert';
 
 class DeleteMeat extends Component {
     static propTypes = {
