+++ frontend/components/elements/ModalWindow.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { CloseX } from './svg/CloseX';
+import { CloseX } from '../svg/CloseX';
 
 const Window = styled.div`
     display: none;
