+++ frontend/components/page_elements/NavBar.js
@@ -1,9 +1,9 @@
 import Link from 'next/link';
 import { Query } from 'react-apollo';
 import styled from 'styled-components';
-import { ALL_CATEGORIES_QUERY } from '../queries/Category';
-import { ALL_MEATS_QUERY } from '../queries/Meat';
-import { DownArrow } from './svg/DownArrow';
+import { ALL_CATEGORIES_QUERY } from '../../queries/Category';
+import { ALL_MEATS_QUERY } from '../../queries/Meat';
+import { DownArrow } from '../svg/DownArrow';
 
 const StyledNavBar = styled.ul`
     float: left;
