--- frontend/mutations/Login.js
@@ -1,13 +1,0 @@
-import gql from 'graphql-tag';
-
-const LOGIN_MUTATION = gql`
-    mutation LOGIN_MUTATION($email: String!, $password: String!) {
-        login(email: $email, password: $password) {
-            id
-            email
-            name
-        }
-    }
-`;
-
-export { LOGIN_MUTATION };
