--- frontend/mutations/Logout.js
@@ -1,11 +1,0 @@
-import gql from 'graphql-tag';
-
-const LOGOUT_MUTATION = gql`
-    mutation LOGOUT_MUTATION {
-        logout {
-            message
-        }
-    }
-`;
-
-export { LOGOUT_MUTATION };
