+++ frontend/pages/login.js
@@ -1,4 +1,4 @@
-import { LoginForm } from '../components/LoginForm';
+import { LoginForm } from '../components/forms/user/LoginForm';
 import { User } from '../components/User';
 import { Homepage } from '../components/Homepage';
