+++ frontend/queries/User.js
@@ -26,5 +26,21 @@ const SINGLE_USER_USERNAME_QUERY = gql`
     }
 `;
 
+const CURRENT_USER_QUERY = gql`
+    query {
+        me {
+            id
+            email
+            name
+            username
+            bio
+            image
+            largeImage
+            permissions
+        }
+    }
+`;
+
 export { SINGLE_USER_QUERY };
 export { SINGLE_USER_USERNAME_QUERY };
+export { CURRENT_USER_QUERY };
